migrate react-streaming test
