+++ test/integration/react-server-components/app/pages/styled-jsx.server.js
