+++ test/integration/react-server-components/test/switchable-runtime.test.js
