+++ test/integration/react-streaming/app/next.config.js
@@ -6,6 +6,6 @@ module.exports = {
   // pageExtensions: ['js', 'ts', 'jsx'], // .tsx won't be treat as page,
   experimental: {
     // serverComponents: true,
-    runtime: 'edge',
+    runtime: 'nodejs',
   },
 }
