+++ test/integration/react-streaming/test/streaming.js
@@ -1,7 +1,11 @@
 /* eslint-env jest */
 import webdriver from 'next-webdriver'
-import { fetchViaHTTP } from 'next-test-utils'
-import { getNodeBySelector } from './utils'
+import { fetchViaHTTP, renderViaHTTP } from 'next-test-utils'
+
+function getNodeBySelector(html, selector) {
+  const $ = cheerio.load(html)
+  return $(selector)
+}
 
 async function resolveStreamResponse(response, onData) {
   let result = ''
@@ -45,55 +49,6 @@ export default function (context, { env, runtime }) {
     expect(content).toMatchInlineSnapshot('"next_streaming_data"')
   })
 
-  it('should support streaming for flight response', async () => {
-    await fetchViaHTTP(context.appPort, '/?__flight__=1').then(
-      async (response) => {
-        const result = await resolveStreamResponse(response)
-        expect(result).toContain('component:index.server')
-      }
-    )
-  })
-
-  it('should support partial hydration with inlined server data', async () => {
-    await fetchViaHTTP(context.appPort, '/partial-hydration', null, {}).then(
-      async (response) => {
-        let gotFallback = false
-        let gotData = false
-        let gotInlinedData = false
-
-        await resolveStreamResponse(response, (_, result) => {
-          gotInlinedData = result.includes('self.__next_s=')
-          gotData = result.includes('next_streaming_data')
-          if (!gotFallback) {
-            gotFallback = result.includes('next_streaming_fallback')
-            if (gotFallback) {
-              expect(gotData).toBe(false)
-              expect(gotInlinedData).toBe(false)
-            }
-          }
-        })
-
-        expect(gotFallback).toBe(true)
-        expect(gotData).toBe(true)
-        expect(gotInlinedData).toBe(true)
-      }
-    )
-
-    // Should end up with "next_streaming_data".
-    const browser = await webdriver(context.appPort, '/partial-hydration')
-    const content = await browser.eval(`window.document.body.innerText`)
-    expect(content).toContain('next_streaming_data')
-
-    // Should support partial hydration: the boundary should still be pending
-    // while another part is hydrated already.
-    expect(await browser.eval(`window.partial_hydration_suspense_result`)).toBe(
-      'next_streaming_fallback'
-    )
-    expect(await browser.eval(`window.partial_hydration_counter_result`)).toBe(
-      'count: 1'
-    )
-  })
-
   it('should not stream to crawlers or google pagerender bot', async () => {
     const res1 = await fetchViaHTTP(
       context.appPort,
@@ -131,5 +86,24 @@ export default function (context, { env, runtime }) {
       expect(res1.headers.get('etag')).toBeDefined()
       expect(res2.headers.get('etag')).toBeDefined()
     }
+  })
+
+  it('should render 500 error correctly', async () => {
+    const errPaths = ['/err', '/err/render']
+    const promises = errPaths.map(async (pagePath) => {
+      const html = await renderViaHTTP(context.appPort, pagePath)
+      if (env === 'dev') {
+        // In dev mode it should show the error popup.
+        expect(html).toContain('Error: oops')
+      } else {
+        expect(html).toContain('custom-500-page')
+      }
+    })
+    await Promise.all(promises)
+  })
+
+  it('should render fallback if error raised from suspense during streaming', async () => {
+    const html = await renderViaHTTP(context.appPort, '/err/suspense')
+    expect(html).toContain('error-fallback')
   })
 }
